Move commands to files
