+++ pyproject.toml
@@ -10,7 +10,7 @@ repository = "https://github.com/fabianhe/pyrepositoryminer"
 documentation = "https://github.com/fabianhe/pyrepositoryminer/blob/master/DOCS.md"
 
 [tool.poetry.scripts]
-pyrepositoryminer = "pyrepositoryminer.main:app"
+pyrepositoryminer = "pyrepositoryminer:app"
 
 [tool.poetry.dependencies]
 python = "^3.8"
