+++ pyrepositoryminer/commands/__init__.py
@@ -1,0 +1,6 @@
+from pyrepositoryminer.commands.analyze import analyze
+from pyrepositoryminer.commands.branch import branch
+from pyrepositoryminer.commands.clone import clone
+from pyrepositoryminer.commands.commits import commits
+
+__all__ = ("analyze", "branch", "clone", "commits")
